+++ pynguin/assertion/noneassertionobserver.py
@@ -15,12 +15,12 @@ import pynguin.testcase.statement as st
 from pynguin.utils.type_utils import is_primitive_type
 
 if TYPE_CHECKING:
-    import pynguin.assertion.outputtrace as ot
+    import pynguin.assertion.statetrace as ot
     import pynguin.testcase.execution as ex
     import pynguin.testcase.variablereference as vr
 
 
-class NoneTraceObserver(ato.AssertionTraceObserver[nte.NoneTraceEntry]):
+class NoneTraceObserver(ato.AssertionTraceObserver):
     """An observer that trace the none-ness of variables."""
 
     def before_statement_execution(
@@ -34,6 +34,8 @@ class NoneTraceObserver(ato.AssertionTraceObserver[nte.NoneTraceEntry]):
         exec_ctx: ex.ExecutionContext,
         exception: Optional[Exception] = None,
     ) -> None:
+        if statement.ret_val is None:
+            return
         if exception is not None:
             return
         if statement.ret_val.is_none_type():
@@ -50,7 +52,7 @@ class NoneAssertionVisitor(st.StatementVisitor):
         self,
         exec_ctx: ex.ExecutionContext,
         variable: vr.VariableReference,
-        trace: ot.OutputTrace[nte.NoneTraceEntry],
+        trace: ot.StateTrace,
     ):
         self._exec_ctx = exec_ctx
         self._variable = variable
@@ -104,19 +106,18 @@ class NoneAssertionVisitor(st.StatementVisitor):
     def visit_assignment_statement(self, stmt) -> None:
         raise NotImplementedError("Assignments are not supported yet")
 
-    def handle(self, statement: st.Statement) -> None:
+    def handle(self, statement: st.VariableCreatingStatement) -> None:
         """Actually handle the given statement.
 
         Args:
             statement: the statement that is visited.
 
         """
-        value = self._exec_ctx.get_variable_value(self._variable)
+        value = self._exec_ctx.get_reference_value(self._variable)
         if is_primitive_type(type(value)):
             return
 
         self._trace.add_entry(
             statement.get_position(),
-            self._variable,
             nte.NoneTraceEntry(self._variable, value is None),
         )