+++ pynguin/assertion/primitiveassertionobserver.py
@@ -15,12 +15,12 @@ import pynguin.testcase.statement as st
 from pynguin.utils.type_utils import is_primitive_type
 
 if TYPE_CHECKING:
-    import pynguin.assertion.outputtrace as ot
+    import pynguin.assertion.statetrace as ot
     import pynguin.testcase.execution as ex
     import pynguin.testcase.variablereference as vr
 
 
-class PrimitiveTraceObserver(ato.AssertionTraceObserver[pte.PrimitiveTraceEntry]):
+class PrimitiveTraceObserver(ato.AssertionTraceObserver):
     """An observer that creates assertions on primitive values."""
 
     def before_statement_execution(
@@ -34,6 +34,8 @@ class PrimitiveTraceObserver(ato.AssertionTraceObserver[pte.PrimitiveTraceEntry]
         exec_ctx: ex.ExecutionContext,
         exception: Optional[Exception] = None,
     ) -> None:
+        if statement.ret_val is None:
+            return
         if exception is not None:
             return
         if statement.ret_val.is_none_type():
@@ -56,7 +58,7 @@ class PrimitiveAssertionVisitor(st.StatementVisitor):
         self,
         exec_ctx: ex.ExecutionContext,
         variable: vr.VariableReference,
-        trace: ot.OutputTrace[pte.PrimitiveTraceEntry],
+        trace: ot.StateTrace,
     ):
         self._exec_ctx = exec_ctx
         self._variable = variable
@@ -110,19 +112,18 @@ class PrimitiveAssertionVisitor(st.StatementVisitor):
     def visit_assignment_statement(self, stmt) -> None:
         raise NotImplementedError("Assignments are not supported yet")
 
-    def handle(self, statement: st.Statement) -> None:
+    def handle(self, statement: st.VariableCreatingStatement) -> None:
         """Actually handle the statement.
 
         Args:
             statement: the statement that is visited.
         """
-        value = self._exec_ctx.get_variable_value(self._variable)
+        value = self._exec_ctx.get_reference_value(self._variable)
         if value is None:
             return
 
         if is_primitive_type(type(value)):
             self._trace.add_entry(
                 statement.get_position(),
-                self._variable,
                 pte.PrimitiveTraceEntry(self._variable, value),
             )