+++ tests/assertion/test_primitiveassertionobserver.py
@@ -22,6 +22,10 @@ import pynguin.assertion.primitivetraceentry as pte
         ("visit_bytes_primitive_statement", 0),
         ("visit_boolean_primitive_statement", 0),
         ("visit_none_statement", 0),
+        ("visit_set_statement", 0),
+        ("visit_list_statement", 0),
+        ("visit_dict_statement", 0),
+        ("visit_tuple_statement", 0),
         ("visit_constructor_statement", 1),
         ("visit_method_statement", 1),
         ("visit_function_statement", 1),
@@ -29,7 +33,7 @@ import pynguin.assertion.primitivetraceentry as pte
 )
 def test_visits(method, call_count):
     exec_ctx = MagicMock()
-    exec_ctx.get_variable_value.return_value = 5
+    exec_ctx.get_reference_value.return_value = 5
     variable = MagicMock()
     trace = MagicMock()
     visitor = pao.PrimitiveAssertionVisitor(exec_ctx, variable, trace)
@@ -48,7 +52,7 @@ def test_visits(method, call_count):
 )
 def test_visits_unimplemented(method):
     exec_ctx = MagicMock()
-    exec_ctx.get_variable_value.return_value = 5
+    exec_ctx.get_reference_value.return_value = 5
     variable = MagicMock()
     trace = MagicMock()
     visitor = pao.PrimitiveAssertionVisitor(exec_ctx, variable, trace)
@@ -60,21 +64,19 @@ def test_visits_unimplemented(method):
 
 def test_handle_primitive():
     exec_ctx = MagicMock()
-    exec_ctx.get_variable_value.return_value = 5
+    exec_ctx.get_reference_value.return_value = 5
     variable = MagicMock()
     trace = MagicMock()
     visitor = pao.PrimitiveAssertionVisitor(exec_ctx, variable, trace)
     statement = MagicMock()
     statement.get_position.return_value = 42
     visitor.handle(statement)
-    trace.add_entry.assert_called_with(
-        42, variable, pte.PrimitiveTraceEntry(variable, 5)
-    )
+    trace.add_entry.assert_called_with(42, pte.PrimitiveTraceEntry(variable, 5))
 
 
 def test_handle_not_primitive():
     exec_ctx = MagicMock()
-    exec_ctx.get_variable_value.return_value = MagicMock()
+    exec_ctx.get_reference_value.return_value = MagicMock()
     variable = MagicMock()
     trace = MagicMock()
     visitor = pao.PrimitiveAssertionVisitor(exec_ctx, variable, trace)
@@ -85,7 +87,7 @@ def test_handle_not_primitive():
 
 def test_handle_none():
     exec_ctx = MagicMock()
-    exec_ctx.get_variable_value.return_value = None
+    exec_ctx.get_reference_value.return_value = None
     variable = MagicMock()
     trace = MagicMock()
     visitor = pao.PrimitiveAssertionVisitor(exec_ctx, variable, trace)